# backend/api.yaml
openapi: 3.0.3
info:
  title: Expense Management API
  description: |
    REST API for expense management system with multi-step approval workflows.
    
    **Features:**
    - JWT-based authentication
    - Role-based access control (Admin, Manager, Employee)
    - Expense submission and tracking
    - Multi-step approval workflows
    - OCR receipt parsing
    - Currency and country integration
  version: 1.0.0
  contact:
    name: API Support
    email: support@expensemanagement.com
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.expensemanagement.com
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /api/expenses:
    post:
      tags:
        - Expenses
      summary: Submit expense
      description: Submit a new expense for approval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitExpenseRequest'
      responses:
        '200':
          description: Expense submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
        '400':
          description: Invalid request
...
  /api/expenses/my/status/{status}:
    get:
      tags:
        - Expenses
      summary: Get expenses by status
      description: Retrieve user's expenses filtered by status
      parameters:
        - name: status
          in: path
          required: true
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED]
      responses:
        '200':
          description: List of expenses with specified status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExpenseResponse'
  /api/expenses/{id}:
    get:
      tags:
        - Expenses
      summary: Get expense by ID
      description: Retrieve a specific expense by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Expense details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
        '404':
          description: Expense not found
  /api/expenses/pending:
    get:
      tags:
        - Expenses
      summary: Get pending expenses
      description: Retrieve all pending expenses for the company (Admin/Manager only)
      responses:
        '200':
          description: List of pending expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExpenseResponse'
        '403':
          description: Forbidden - requires Admin or Manager role
...
  /api/approvals/expense/{expenseId}:
    get:
      tags:
        - Approvals
      summary: Get approval steps for expense
      description: Retrieve all approval steps for a specific expense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of approval steps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApprovalStepResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SubmitExpenseRequest:
      type: object
      required:
        - amount
        - currency
        - category
        - expenseDate
      properties:
        amount:
          type: number
          format: decimal
          example: 150.00
        currency:
          type: string
          example: USD
        category:
          type: string
          example: Office Supplies
        description:
          type: string
          example: Laptop accessories
        expenseDate:
          type: string
          format: date
          example: 2025-10-01
        receiptImageUrl:
          type: string
          nullable: true
          example: https://storage.example.com/receipts/123.jpg
    ExpenseResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        userFullName:
          type: string
        amount:
          type: number
          format: decimal
        currency:
          type: string
        category:
          type: string
        description:
          type: string
        expenseDate:
          type: string
          format: date
        receiptImageUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        createdAt:
          type: string
          format: date-time
...
    ApprovalStepResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        expenseId:
          type: integer
          format: int64
        approverId:
          type: integer
          format: int64
        approverName:
          type: string
        sequence:
          type: integer
        decision:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        comments:
          type: string
          nullable: true
        decidedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...